Rename before fixing casing
